requirements.txt
pymilvus==2.3.3
requests==2.31.0
python-dotenv==1.0.0
flask==3.0.0
flask-cors==6.0.2
sentence-transformers
PyPDF2==3.0.1
anthropic
beautifulsoup4
lxml
spacy>=3.7.0
textblob>=0.17.1
networkx>=3.2
pandas>=2.1.0
matplotlib>=3.8.0
wordcloud>=1.9.0
reportlab>=4.0.0
Pillow>=10.0.0
